primitives/beefy: move Payload to its own file
